Add "finish-args-flatpak-spawn-access" exception for com.dec05eba.gpu_screen_recorder
The app requires flatpak-spawn to capture a monitor on AMD/Intel GPUs.
It runs flatpak-spawn pkexec to run an external program as root
(the program is part of the flatpak), because AMD/Intel requires
root access to capture a monitor (KMS).

This is done to restrict root access to a very small executable
that only does one thing (access to KMS), as opposed to running the
whole flatpak (with network access) as root.