Add xdg-config for org.gimp.GIMP.
This permission was initially added for our use case (or at least
partly, maybe others were asking too).
See: https://github.com/flatpak/flatpak/issues/328#issuecomment-260565160

We are actually considering dropping this for GIMP 3. But for the 2.10
series and the 2.99 development series, we need to keep continuity. We
cannot suddenly break everyone's configuration in the middle of a
release series.