Add a check to find if exportable icons exist
This goes through icons installed in `/app/share/icons/hicolor` and
checks if any icons matching appid patterns that flatpak exports are
installed. If not return an error because without any icons exported
the icon referenced in desktop file won't be found and appear as empty
in DE overview.

This is seperate from the icon generated by appstream in app-info
because

- appinfo icons are not used by desktop files and flatpak does not
  export them
- it may happen that appstream compose created an icon and named it
  using FLATPAK_ID when no such icon with the correct filename was
  actually installed. The end result is, the flatpak ends up with no
  exportable icons