exceptions: allow automerge for com.fightcade.Fightcade
The Fightcade client needs to be in sync with the current server version
to function. If an update causes breakages the end result is identical
to having an outdated Flatpak (the client will not work in either case.)

Allow Flathubbot automerging to keep the client in sync even when a
developer with merge access is unavailable.