appid: Exclude Devel suffixes from domain check
when an extra devel suffix gets added to the original appid, the
rDNS check will take in all components of the appid but their might not
be a webpage available there. So standardise on the `Devel` suffix
and skip on that.

Devel is not to be allowed on Flathub submission, submission reviews
need to take note of that