Add linter exception for ~/.config/.X-AIR-Edit
The application does not respect $XDG_CONFIG_HOME and always uses the
path ~/.config/.X-AIR-Edit (including the second dot) as the directory
to put its configuration files in.

See <https://github.com/flathub/com.behringer.XAirEdit/pull/20> for a
discussion about this.

We have asked the vendor to fix this behavior in
<https://ideas.behringer.com/p/x-air-edit-linux-version-does-not-respect-xdg_config_home>,
but in the mean time, we'd like an exception for this so we no longer
have to grant the application access to the whole home directory.