cli: Add a way to specify local exceptions to ignore
this seems a common pattern for people running linter as part of their
upstream CI. Not every error we emit for Flathub is relevant for their
CI, and we do not need to accept exceptions for errors related to
upstream CI eg. if the ID is .Devel and not on Flathub.

So `--user-exceptions` will allow them to specify a baseline of errors
to ignore for use in CI