Separate metainfo check and run validation for all refs
We need to run some kind of basic metainfo validation if it is present.
The previous code was skipping the entire `_validate()` function if
it was a baseapp or non-application.

This meant runtimes (extensions) can bypass the checks and send
broken metainfo to be composed in catalogue.

So now, we skip only the "appstream-metainfo-missing" for baseapps
and non-applications and otherwise run `appstreamcli validate` on the
input metainfo file if it is present.

This also extends appstreamcli validate to metainfo files installed
by subnames of appid as Flatpak can collect these and put in the
catalogue although "appstream-metainfo-missing" is raised only if
the metainfo matching the appid exactly is not found.

Since an app can export multiple extensions all of which are runtime
refs we have to run the repo check in metainfo for all refs.