Support checking multiple primary refs sharing same ref id
A repo can contain more than one "app" ref, we should be checking
all of them not just the first one as the content in them might differ
in atypical setups.

This also helps to reduce an extra linter run on "per arch" repos. The
assumption is that multiple app refs will share the same ref id

[release]