Parse required-flatpak and enforce a minimum version for new metadata
Current Flatpak implementation does not provide any backwards compat
for these metadata, so if a maintainer sets it in the manifest and
the user doesn't have a recent enough Flatpak it will be effectively
equivalent to not having the permission at all for the user thus
breaking the intended use of the application.

So enforce a minimum required-flatpak for these new contetx metadata.
1.16.0 was the first "stable" Flatpak release to have this, don't
consider unstable releases.

To be removed when backwards compat is figured out in https://github.com/flatpak/flatpak/issues/5681