Exclude home access from linter checks for GIMP
Exclude `home` access in finish-args from failing in GIMP project. Currently GIMP allows `host` access which is an overkill and insecure in terms of sandboxing. The linter does currently  allow `host` access as the permission predates linter rule. 

I will remove that later - if and when the PR to switch `host` to `home` gets accepted and merged in org.gimp.GIMP flathub repo.