exceptions: Enable automerge for KDE Games
Allow `automerge-flathubbot-prs` for all KDE Games to let the bot merge
the PRs automatically for those projects.

KDE Games are low risk, low impact apps that are unlikely to break.
Enabling auto-updates for those apps will free-up some time for the KDE
team to look at the other apps which are more important.